General refactor plus Age and Season Services

+++ backend/src/main/java/com/raulmvp/dao/UniverseDao.java
@@ -10,36 +10,52 @@ import java.util.List;
 @Repository
 public class UniverseDao extends AbstractDao {
 
+    private static final String PARAMETER_NAME = "name";
+    private static final String PARAMETER_USERNAME = "username";
+
     public UniverseEntity createUniverse(UniverseEntity universeEntity) {
         em.persist(universeEntity);
         return universeEntity;
     }
 
-    public UniverseEntity getUniverseById(int id) {
-        return em.find(UniverseEntity.class, id);
+    public UniverseEntity getUniverseById(int aId) {
+        return em.find(UniverseEntity.class, aId);
     }
 
-    public List<UniverseEntity> getUniverseListByUsername(String username) {
-        List<UniverseEntity> uLst = new ArrayList<>();
+    public List<UniverseEntity> getUniverseListByUsername(String aUsername) {
+
         Query q = em.createNamedQuery("UniverseEntity.getByUsername");
-        q.setParameter("username", username);
+        q.setParameter(PARAMETER_USERNAME, aUsername);
 
-        uLst = q.getResultList();
-        return uLst;
+        return q.getResultList();
     }
 
-    public UniverseEntity getUniverseByName(String name) {
+    public UniverseEntity getUniverseByName(String aName) {
 
         Query q = em.createNamedQuery("UniverseEntity.getByName");
-        q.setParameter("name", name);
+        q.setParameter(PARAMETER_NAME, aName);
 
         UniverseEntity universeEntity = (UniverseEntity) q.getSingleResult();
 
         return universeEntity;
     }
 
-    public void deleteUniverse(String name) {
-        UniverseEntity universeEntity = getUniverseByName(name);
-        em.remove(universeEntity);
+    public Boolean deleteUniverse(String aName) {
+
+        UniverseEntity universeEntity = getUniverseByName(aName);
+        Boolean isDeleted = Boolean.FALSE;
+
+        if (universeEntity == null) {
+            return isDeleted;
+        }
+
+        try {
+            em.remove(universeEntity);
+            isDeleted = Boolean.TRUE;
+        } catch (IllegalArgumentException e) {
+            System.out.println("UniverseDao - Delete: " + e);
+        }
+
+        return isDeleted;
     }
 }
